+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## ✈ the latest
 - `squirls-cli`
   - allow to block-compress (`bgzip`) the annotated VCF on the fly by using `-f vcfgz` option
-  - do not fail due to an exception being thrown by Jannovar when annotating a VCF
+  - do not fail due to an exception being thrown by Jannovar when annotating a VCF 
+  - do not sort variants in the VCF output if the required header lines are missing
 - `squirls-ingest`
   - prevent ingesting transcripts with duplicate *ENST* id